Back-end/requirements.txt
3.10.0
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==2.0.0
pydantic==2.0.0
python-dotenv==1.0.0
motor
py -3.10 -m venv venv_mediapipe

.\venv_mediapipe\Scripts\activate

pip install mediapipe opencv-python